requirements.txt
pandas
numpy
matplotlib
xgboost
scikit-learn
rich
pyyaml
